GUI/requirements_GUI.txt
customtkinter==5.2.2
darkdetect==0.8.0
numpy==2.2.3
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
